Improve C API by handling C++ exceptions
The InfluxDB class uses C++ exceptions to let the caller of its own methods
know when something goes wrong. Since exceptions are not managed in C, we need
to handle them directly in the C++ code and return meaningful error code back
to the C interface.

Improve the C interface by returning meaningful error code when exceptions are
handled in InfluxDB C++ code.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>